restructured code to use APIManager singleton class. Now, I am able to display the song title of a queried song. Doesn't load new data however.
