+++ Project-Hotpot/Project-Hotpot/APIManager.swift
@@ -1,0 +1,209 @@
+//
+//  APIManager.swift
+//  Project-Hotpot
+//
+//  Created by Harini Sundaram on 7/11/22.
+//
+
+import Foundation
+import UIKit
+
+class APIManager: NSObject {
+//    @IBOutlet weak var songTitle: UILabel!
+    var curr_song_label:String?
+    
+//    class NetworkManager {
+
+        // MARK: - Properties
+
+        private static var sharedAPIManager: APIManager = {
+            let SpotifyManager = APIManager()
+            return SpotifyManager
+        }()
+
+        // MARK: - Initializers
+        //doing this way so that only 1 instance can be created
+        override private init() {
+            NSLog("API Manager Initialized")
+        }
+
+        // MARK: - Accessors
+        
+        class func shared() -> APIManager {
+            return sharedAPIManager
+        }
+
+    
+    var lastPlayerState: SPTAppRemotePlayerState?
+    // MARK: - Spotify Authorization & Configuration
+    var responseCode: String? {
+        didSet {
+            fetchAccessToken { (dictionary, error) in
+                if let error = error {
+                    print("Fetching token request error \(error)")
+                    return
+                }
+                let accessToken = dictionary!["access_token"] as! String
+                DispatchQueue.main.async {
+                    self.appRemote.connectionParameters.accessToken = accessToken
+                    self.appRemote.connect()
+                }
+            }
+        }
+    }
+
+    lazy var appRemote: SPTAppRemote = {
+        let appRemote = SPTAppRemote(configuration: configuration, logLevel: .debug)
+        appRemote.connectionParameters.accessToken = self.accessToken
+        appRemote.delegate = self
+        return appRemote
+    }()
+
+    var accessToken = UserDefaults.standard.string(forKey: accessTokenKey) {
+        didSet {
+            let defaults = UserDefaults.standard
+            defaults.set(accessToken, forKey: accessTokenKey)
+        }
+    }
+
+    lazy var configuration: SPTConfiguration = {
+        let configuration = SPTConfiguration(clientID: spotifyClientId, redirectURL: redirectUri)
+//
+
+        // Set the playURI to a non-nil value so that Spotify plays music after authenticating
+        // otherwise another app switch will be required
+        // for now, leaving this line commented out purposefully
+        // configuration.playURI = ""
+        // Set these url's to your backend which contains the secret to exchange for an access token
+        // You can use the provided ruby script spotify_token_swap.rb for testing purposes
+        configuration.tokenSwapURL = URL(string: "http://localhost:1234/swap")
+        configuration.tokenRefreshURL = URL(string: "http://localhost:1234/refresh")
+        return configuration
+    }()
+    
+//    static let spotifyPlayer = SPTAudioStreamingController(clientId: spotifyClientId)
+
+    lazy var sessionManager: SPTSessionManager? = {
+        let manager = SPTSessionManager(configuration: configuration, delegate: self)
+        return manager
+    }()
+
+    func update(playerState: SPTAppRemotePlayerState) {
+        NSLog("updated!")
+        self.lastPlayerState = playerState
+        self.curr_song_label = playerState.track.name
+    }
+}
+
+
+
+// MARK: - SPTAppRemoteDelegate
+extension APIManager: SPTAppRemoteDelegate {
+    func appRemoteDidEstablishConnection(_ appRemote: SPTAppRemote) {
+        appRemote.playerAPI?.delegate = self
+        appRemote.playerAPI?.subscribe(toPlayerState: { (success, error) in
+            if let error = error {
+                print("Error subscribing to player state:" + error.localizedDescription)
+            }
+        })
+        fetchPlayerState()
+    }
+
+    func appRemote(_ appRemote: SPTAppRemote, didDisconnectWithError error: Error?) {
+        lastPlayerState = nil
+        
+    }
+
+    func appRemote(_ appRemote: SPTAppRemote, didFailConnectionAttemptWithError error: Error?) {
+        lastPlayerState = nil
+    }
+}
+
+// MARK: - SPTAppRemotePlayerAPIDelegate
+extension APIManager: SPTAppRemotePlayerStateDelegate {
+    func playerStateDidChange(_ playerState: SPTAppRemotePlayerState) {
+        debugPrint("Spotify Track name: %@", playerState.track.name)
+        update(playerState: playerState)
+    }
+}
+
+// MARK: - SPTSessionManagerDelegate
+ extension APIManager: SPTSessionManagerDelegate {
+    func sessionManager(manager: SPTSessionManager, didFailWith error: Error) {
+        if error.localizedDescription == "The operation couldn’t be completed. (com.spotify.sdk.login error 1.)" {
+            print("AUTHENTICATE with WEBAPI")
+        }
+    }
+
+    func sessionManager(manager: SPTSessionManager, didRenew session: SPTSession) {
+    }
+
+    func sessionManager(manager: SPTSessionManager, didInitiate session: SPTSession) {
+        appRemote.connectionParameters.accessToken = session.accessToken
+        appRemote.connect()
+        NSLog("session connected")
+    }
+}
+
+// MARK: - Networking
+extension APIManager {
+
+    func fetchAccessToken(completion: @escaping ([String: Any]?, Error?) -> Void) {
+        let url = URL(string: "https://accounts.spotify.com/api/token")!
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        let spotifyAuthKey = "Basic \((spotifyClientId + ":" + spotifyClientSecretKey).data(using: .utf8)!.base64EncodedString())"
+        request.allHTTPHeaderFields = ["Authorization": spotifyAuthKey,
+                                       "Content-Type": "application/x-www-form-urlencoded"]
+
+        var requestBodyComponents = URLComponents()
+        let scopeAsString = stringScopes.joined(separator: " ")
+
+        requestBodyComponents.queryItems = [
+            URLQueryItem(name: "client_id", value: spotifyClientId),
+            URLQueryItem(name: "grant_type", value: "authorization_code"),
+            URLQueryItem(name: "code", value: responseCode!),
+            URLQueryItem(name: "redirect_uri", value: redirectUri.absoluteString),
+            URLQueryItem(name: "code_verifier", value: ""), // not currently used
+            URLQueryItem(name: "scope", value: scopeAsString),
+        ]
+
+        request.httpBody = requestBodyComponents.query?.data(using: .utf8)
+
+        let task = URLSession.shared.dataTask(with: request) { data, response, error in
+            guard let data = data,                              // is there data
+                  let response = response as? HTTPURLResponse,  // is there HTTP response
+                  (200 ..< 300) ~= response.statusCode,         // is statusCode 2XX
+                  error == nil else {                           // was there no error, otherwise ...
+                      print("Error fetching token \(error?.localizedDescription ?? "")")
+                      return completion(nil, error)
+                  }
+            let responseObject = try? JSONSerialization.jsonObject(with: data) as? [String: Any]
+            print("Access Token Dictionary=", responseObject ?? "")
+            completion(responseObject, nil)
+        }
+        task.resume()
+    }
+
+    func fetchArtwork(for track: SPTAppRemoteTrack) {
+        appRemote.imageAPI?.fetchImage(forItem: track, with: CGSize.zero, callback: { [weak self] (image, error) in
+            if let error = error {
+                print("Error fetching track image: " + error.localizedDescription)
+            }
+        })
+    }
+
+    func fetchPlayerState() {
+        NSLog("in fetch player")
+        appRemote.playerAPI?.getPlayerState({ [weak self] (playerState, error) in
+            if let error = error {
+                print("Error getting player state:" + error.localizedDescription)
+            } else if let playerState = playerState as? SPTAppRemotePlayerState {
+                self?.update(playerState: playerState)
+
+            }
+        })
+    }
+}
+
+
